// src/app/demo/models/modelsgeneral/estadoCivilviewmodel.ts
export class EstadoCivil{
    codigo?: number;
    civi_Id?: number;
    civi_Descripcion?:string;
    usua_Creacion?:number;
    usua_Modificacion?: number;


    civi_FechaCreacion?: string;
    civi_FechaModificacion?: string;
    civi_Estado?: boolean;


    usuaCreacion?:string;
    usuaModificacion?:string;
}

export class DropDownEstadoCivil{
    civi_Id?: number;
    civi_Descripcion?:string;
}